# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "OpenVPN Access Server Management API with Bulk Operations, Advanced Search, and Comprehensive Redis Caching",
        "title": "GoVPN API Enhanced with Advanced Redis Caching",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/openvpn/bulk/groups/actions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Perform actions on multiple groups (enable/disable)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Bulk group actions",
                "parameters": [
                    {
                        "description": "Bulk group actions data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkGroupActionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BulkGroupActionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/bulk/groups/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create multiple groups at once",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Bulk create groups",
                "parameters": [
                    {
                        "description": "Bulk group creation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkCreateGroupsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.BulkCreateGroupsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/bulk/groups/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Import groups from CSV, JSON, or XLSX file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Import groups from file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Groups file (CSV/JSON/XLSX)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "csv",
                            "json",
                            "xlsx"
                        ],
                        "type": "string",
                        "description": "File format",
                        "name": "format",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Dry run mode (validate only)",
                        "name": "dryRun",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Override existing groups",
                        "name": "override",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ImportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/bulk/groups/template": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Download template file for group import",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Export group template",
                "parameters": [
                    {
                        "enum": [
                            "csv",
                            "xlsx"
                        ],
                        "type": "string",
                        "default": "csv",
                        "description": "Template format",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group template file",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/openvpn/bulk/users/actions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Perform actions on multiple users (enable/disable/reset-otp)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Bulk user actions",
                "parameters": [
                    {
                        "description": "Bulk user actions data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkUserActionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BulkActionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/bulk/users/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create multiple users at once",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Bulk create users",
                "parameters": [
                    {
                        "description": "Bulk user creation data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkCreateUsersRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.BulkCreateUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request too large",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/openvpn/bulk/users/extend": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Extend expiration date for multiple users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Bulk extend user expiration",
                "parameters": [
                    {
                        "description": "Bulk user extension data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkUserExtendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BulkActionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/bulk/users/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Import users from CSV, JSON, or XLSX file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Import users from file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Users file (CSV/JSON/XLSX)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "csv",
                            "json",
                            "xlsx"
                        ],
                        "type": "string",
                        "description": "File format",
                        "name": "format",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Dry run mode (validate only)",
                        "name": "dryRun",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Override existing users",
                        "name": "override",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ImportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "File too large",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/bulk/users/template": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Download template file for user import",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Bulk Operations"
                ],
                "summary": "Export user template",
                "parameters": [
                    {
                        "enum": [
                            "csv",
                            "xlsx"
                        ],
                        "type": "string",
                        "default": "csv",
                        "description": "Template format",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User template file",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BulkGroupActionResponse": {
            "type": "object",
            "properties": {
                "failed": {
                    "type": "integer",
                    "example": 1
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.BulkGroupOperationResult"
                    }
                },
                "success": {
                    "type": "integer",
                    "example": 4
                },
                "total": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "dto.BulkGroupActionsRequest": {
            "type": "object",
            "required": [
                "action",
                "groupNames"
            ],
            "properties": {
                "action": {
                    "type": "string",
                    "enum": [
                        "enable",
                        "disable"
                    ]
                },
                "groupNames": {
                    "type": "array",
                    "maxItems": 50,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and RSA256 JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}